requirements.txt
pandas==2.2.3
opencv-python==4.10.0.84
tqdm==4.67.1
requests==2.32.3
pillow==11.0.0
emnist==0.0
matplotlib==3.10.1